Transit Forms updated to use typeahead-based vendor solution
Previously, when setting vendor for an asset, the Vendor lookup table had to be populated, which in most cases it wasn't.  By adding a virtual attribute for vendor_name, we allow the user to add a vendor by name- specifically, we use find_or_create_by to build the vendor record if it doesn't exist.  The forms updated here now use typeahead and a virtual attribute to build vendors on the fly as necessary