[Fixes #119226523] Fixing up the upload sheet tempalte to match columns names with object parameter names.
Updates the substitution used on the fields. It had been gsub! to gsub. The ! changes the string in place which means later records couldn't match on the original string and would break.

Also added a fix to have the land_owner_type convert the string into an fta_owner_type.