bug fix (priority 3)
2.3.0 QA Bug List
2.3.0 QA Bug List
100%
10

When adding a new facility you choose FTA Class Type and the Facility Categorization before going to the second page for entering fields. If the user selects Component or Subcomponent in the Facility Categorization, that value should be reflected by the radio button in the second page.
Screen reader support enabled.

When adding a new facility you choose FTA Class Type and the Facility Categorization before going to the second page for entering fields. If the user selects Component or Subcomponent in the Facility Categorization, that value should be reflected by the radio button in the second page.